requirements.txt
alembic==1.11.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==7.1.2
cryptography==41.0.2
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.0
Flask-User==1.0.1.5
Flask-WTF==0.14.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.0
Mako==1.2.4
MarkupSafe==2.0.0rc2
passlib==1.7.4
protobuf==3.20.3
psycopg2==2.9.6
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.0.0
wikipedia==1.4.0
WTForms==3.0.1
